Fin del proyecto

+++ app/api/todos/route.ts
@@ -1,3 +1,4 @@
+import { getUserSessionServer } from '@/auth/actions/auth-actions';
 import prisma from '@/lib/prisma'
 import { NextResponse } from 'next/server'
 import * as yup from 'yup';
@@ -33,10 +34,14 @@ const postSchema = yup.object({
 
 export async function POST(request: Request) {
 
+    const user = await getUserSessionServer();
+
+    if (!user) return NextResponse.json('No autorizado', {status: 401 });
+
     try {
         const { complete, description } = await postSchema.validate(await request.json());
 
-        const todo = await prisma.todo.create({ data: { complete, description } })
+        const todo = await prisma.todo.create({ data: { complete, description, userId: user.id } })
 
         return NextResponse.json(todo);
 
@@ -49,8 +54,12 @@ export async function POST(request: Request) {
 
 export async function DELETE() {
 
+    const user = await getUserSessionServer();
+
+    if (!user) return NextResponse.json('No autorizado', {status: 401 });
+
     try {
-        const deletedTodos = await prisma.todo.deleteMany({where: {complete: true}})
+        const deletedTodos = await prisma.todo.deleteMany({where: {complete: true, userId: user.id}})
 
         if( deletedTodos.count > 0 ) {
             return NextResponse.json({message: 'Todos completados eliminados'});
